Avoid inserting initial lock if there is already a lock in the table.
There is currently a race in ensureTable where the end result could
be that multiple rows are inserted into the locks table and putting
the lock table in a non-recoverable state where the no lock can later
be acquired. The reason for race is that the number of locks in the
lock table is first retrieved and a new row is inserted if that number
is zero. If a new lock is inserted right after the number of locks was
first retrived, another lock will be inserted.

This is fixed by using insert into select where no row is inserted if
there is already a row in table, which is the desired end state.